Extract separate components from FeaturesGrid
