+++ packages/plans-grid-next/src/components/features-grid/billing-timeframes.tsx
@@ -1,0 +1,35 @@
+import classNames from 'classnames';
+import { GridPlan } from '../../types';
+import PlanDivOrTdContainer from '../plan-div-td-container';
+import BillingTimeframe from '../shared/billing-timeframe';
+
+type BillingTimeframesProps = {
+	renderedGridPlans: GridPlan[];
+	showRefundPeriod?: boolean;
+	options?: {
+		isTableCell?: boolean;
+	};
+};
+
+const BillingTimeframes = ( props: BillingTimeframesProps ) => {
+	const { options, renderedGridPlans, showRefundPeriod } = props;
+
+	return renderedGridPlans.map( ( { planSlug } ) => {
+		const classes = classNames(
+			'plan-features-2023-grid__table-item',
+			'plan-features-2023-grid__header-billing-info'
+		);
+
+		return (
+			<PlanDivOrTdContainer
+				className={ classes }
+				isTableCell={ options?.isTableCell }
+				key={ planSlug }
+			>
+				<BillingTimeframe planSlug={ planSlug } showRefundPeriod={ showRefundPeriod } />
+			</PlanDivOrTdContainer>
+		);
+	} );
+};
+
+export default BillingTimeframes;
