+++ packages/plans-grid-next/src/components/features-grid/plan-headers.tsx
@@ -1,0 +1,33 @@
+import { getPlanClass } from '@automattic/calypso-products';
+import classNames from 'classnames';
+import { GridPlan } from '../../types';
+import PlanDivOrTdContainer from '../plan-div-td-container';
+
+type PlanHeadersProps = {
+	renderedGridPlans: GridPlan[];
+	options?: {
+		isTableCell?: boolean;
+	};
+};
+
+const PlanHeaders = ( props: PlanHeadersProps ) => {
+	const { options, renderedGridPlans } = props;
+
+	return renderedGridPlans.map( ( { planSlug, planTitle } ) => {
+		const headerClasses = classNames( 'plan-features-2023-grid__header', getPlanClass( planSlug ) );
+
+		return (
+			<PlanDivOrTdContainer
+				key={ planSlug }
+				className="plan-features-2023-grid__table-item"
+				isTableCell={ options?.isTableCell }
+			>
+				<header className={ headerClasses }>
+					<h4 className="plan-features-2023-grid__header-title">{ planTitle }</h4>
+				</header>
+			</PlanDivOrTdContainer>
+		);
+	} );
+};
+
+export default PlanHeaders;
