+++ packages/plans-grid-next/src/components/features-grid/plan-logos.tsx
@@ -1,0 +1,29 @@
+import { GridPlan } from '../../types';
+import PlanLogo from '../plan-logo';
+
+type PlanLogosProps = {
+	isInSignup: boolean;
+	renderedGridPlans: GridPlan[];
+	options?: {
+		isTableCell?: boolean;
+	};
+};
+
+const PlanLogos = ( props: PlanLogosProps ) => {
+	const { isInSignup, options, renderedGridPlans } = props;
+
+	return renderedGridPlans.map( ( { planSlug }, index ) => {
+		return (
+			<PlanLogo
+				key={ planSlug }
+				planSlug={ planSlug }
+				planIndex={ index }
+				renderedGridPlans={ renderedGridPlans }
+				isInSignup={ isInSignup }
+				isTableCell={ options?.isTableCell }
+			/>
+		);
+	} );
+};
+
+export default PlanLogos;
