+++ packages/plans-grid-next/src/components/features-grid/plan-price.tsx
@@ -1,0 +1,36 @@
+import { GridPlan } from '../../types';
+import PlanFeatures2023GridHeaderPrice from '../header-price';
+import PlanDivOrTdContainer from '../plan-div-td-container';
+
+type PlanPriceProps = {
+	currentSitePlanSlug?: string | null;
+	planUpgradeCreditsApplicable?: number | null;
+	renderedGridPlans: GridPlan[];
+	options?: {
+		isTableCell?: boolean;
+	};
+};
+
+const PlanPrice = ( props: PlanPriceProps ) => {
+	const { currentSitePlanSlug, options, planUpgradeCreditsApplicable, renderedGridPlans } = props;
+
+	return renderedGridPlans.map( ( { planSlug } ) => {
+		return (
+			<PlanDivOrTdContainer
+				scope="col"
+				key={ planSlug }
+				className="plan-features-2023-grid__table-item plan-price"
+				isTableCell={ options?.isTableCell }
+			>
+				<PlanFeatures2023GridHeaderPrice
+					planSlug={ planSlug }
+					planUpgradeCreditsApplicable={ planUpgradeCreditsApplicable }
+					currentSitePlanSlug={ currentSitePlanSlug }
+					visibleGridPlans={ renderedGridPlans }
+				/>
+			</PlanDivOrTdContainer>
+		);
+	} );
+};
+
+export default PlanPrice;
