+++ packages/plans-grid-next/src/components/features-grid/plan-tagline.tsx
@@ -1,0 +1,27 @@
+import { GridPlan } from '../../types';
+import PlanDivOrTdContainer from '../plan-div-td-container';
+
+type PlanTaglineProps = {
+	renderedGridPlans: GridPlan[];
+	options?: {
+		isTableCell?: boolean;
+	};
+};
+
+const PlanTagline = ( props: PlanTaglineProps ) => {
+	const { options, renderedGridPlans } = props;
+
+	return renderedGridPlans.map( ( { planSlug, tagline } ) => {
+		return (
+			<PlanDivOrTdContainer
+				key={ planSlug }
+				className="plan-features-2023-grid__table-item"
+				isTableCell={ options?.isTableCell }
+			>
+				<div className="plan-features-2023-grid__header-tagline">{ tagline }</div>
+			</PlanDivOrTdContainer>
+		);
+	} );
+};
+
+export default PlanTagline;
